Refactor predicted_ispu columns to use decimal type for accurate value representation in predictions and categories

+++ database/migrations/2025_09_07_122806_add_full_prediction_columns_to_predict_iaqi_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('predicted_category', 100)->change();
 
             // ISPU RI
-            $table->unsignedSmallInteger('predicted_ispu')->nullable()->after('predicted_aqi');
+            $table->decimal('predicted_ispu', 5, 2)->nullable()->after('predicted_aqi');
             $table->string('predicted_category_ispu', 100)->nullable()->after('predicted_ispu');
 
             // Metrics & model info
